+++ include/NeoN/finiteVolume/cellCentred/boundary/volume/extrapolated.hpp
@@ -77,6 +77,8 @@ public:
 
     static std::string name() { return "extrapolated"; }
 
+    std::string getName() const override { return name(); }
+
     static std::string doc() { return "TBD"; }
 
     static std::string schema() { return "none"; }