+++ src/finiteVolume/cellCentred/boundary/volume/processor.cpp
@@ -1,0 +1,48 @@
+// SPDX-FileCopyrightText: 2023 - 2026 NeoN authors
+//
+// SPDX-License-Identifier: MIT
+
+#include "NeoN/core/parallelAlgorithms.hpp"
+#include "NeoN/core/primitives/scalar.hpp"
+#include "NeoN/core/primitives/vec3.hpp"
+#include "NeoN/finiteVolume/cellCentred/boundary/volume/processor.hpp"
+
+namespace NeoN::finiteVolume::cellCentred::volumeBoundary::detail
+{
+
+template<typename ValueType>
+void setProcBoundaryValue(
+    Field<ValueType>& domainVector,
+    const UnstructuredMesh& mesh,
+    std::pair<localIdx, localIdx> range
+)
+{
+    const auto iVector = domainVector.internalVector().view();
+
+    auto [refGradient, value, valueFraction, refValue, faceCells] = views(
+        domainVector.boundaryData().refGrad(),
+        domainVector.boundaryData().value(),
+        domainVector.boundaryData().valueFraction(),
+        domainVector.boundaryData().refValue(),
+        mesh.boundaryMesh().faceOwners()
+    );
+
+    NeoN::parallelFor(
+        domainVector.exec(),
+        range,
+        NEON_LAMBDA(const localIdx i) {
+            refGradient[i] = zero<ValueType>();
+            value[i] = iVector[faceCells[i]];
+            valueFraction[i] = 0.0;
+            refValue[i] = zero<ValueType>();
+        },
+        "setProcBoundaryValue"
+    );
+}
+
+template void setProcBoundaryValue<
+    scalar>(Field<scalar>&, const UnstructuredMesh&, std::pair<localIdx, localIdx>);
+template void
+setProcBoundaryValue<Vec3>(Field<Vec3>&, const UnstructuredMesh&, std::pair<localIdx, localIdx>);
+
+} // namespace NeoN::finiteVolume::cellCentred::volumeBoundary::detail